moved most cache calculation to user code
